requirements.txt
Click==7.0
pip-tools==3.1.0
PyYAML==3.13
six==1.11.0
